+++ app/Http/Controllers/TransferDemandController.php
@@ -262,10 +262,19 @@ class TransferDemandController extends Controller
         $transferDemand = TransferDemand::findOrFail($id);
         // $productsTransfersDemandsLines = $transferDemand ? $transferDemand->productsTransfersDemandsLines : null;
         try {
-            $transferDemand->delete();
+            $success = false;
+            $message = "";
+            if (empty($transferDemand->transfers) || sizeof($transferDemand->transfers) == 0 && empty($transferDemand->productsTransfersDemandsLines) || sizeof($transferDemand->productsTransfersDemandsLines) == 0) {
+                // dd('delete');
+                $transferDemand->delete();
 
-            $success = true;
-            $message = "Suppression effectuée avec succès.";
+                $success = true;
+                $message = "Suppression effectuée avec succès.";
+            } else {
+                // dd('not delete');
+                $message = "Cette demande de transfert ne peut être supprimée car elle a servi dans des traitements.";
+            }
+            
             return new JsonResponse([
                 'transferDemand' => $transferDemand,
                 'success' => $success,
